### 2018-07-26 #851 @whedon
[PRE REVIEW]: logKDE: log-transformed kernel density estimation
**Submitting author:** @hiendn (<a href="http://orcid.org/0000-0002-9958-432X">Hien Nguyen</a>)
**Repository:** <a href="https://github.com/andrewthomasjones/logKDE" target ="_blank">https://github.com/andrewthomasjones/logKDE</a>
**Version:** v0.3.1
