### 2018-07-27 078c93a
Update paper.md
