### 2018-07-27 1247167
Minor change to readme
