+++ README.Rmd
@@ -181,7 +181,7 @@ In this example, we showcase the variety of plugin bandwidth estimators that are
 library(logKDE)
 
 ## Set a random seed.
-set.seed(10)
+set.seed(1)
 
 ## Generate strictly positive data.
 ## Data are generated from a chi-squared distribution with 12 degrees of freedom.
