### 2018-08-04 #851 @arfon
:wave: @sctyner, @AlexAxthelm, @corybrunson - would you be willing to review this submission for JOSS?
