### 2018-08-04 #851 @corybrunson
I should not volunteer before emptying my current queue; i will make a note to check back once i have.

(By the way, i don't recall getting a notification 9 days ago from the comment by @hiendn . Should i have?)
