### 2018-08-04 #851 @arfon
> (By the way, i don't recall getting a notification 9 days ago from the comment by @hiendn . Should i have?)

I *think* so, not sure why you didn't...
