### 2018-08-04 #851 @arfon
> @arfon then i may have overlooked it; if so, i apologize, and i'll make a point to pay closer attention. Thanks!

Don't worry about it at all! GitHub notifications can be hard to deal with :grin:
