### 2018-08-04 #851 @arfon
@strengejacke @hiendn - see you both over in #870 for the actual review.
