### 2018-08-05 #870 @hiendn
Hi @arfon and @strengejacke. I'm looking forward to working with you both! Do not hesitate to forward any queries to me or to my coauthor @andrewthomasjones, who is the owner of the GitHub repository.
