### 2018-08-05 andrewthomasjones/logKDE#1 @strengejacke
[JOSS-review] add DOI to references in paper-bib, where possible
This one is related to https://github.com/openjournals/joss-reviews/issues/870.

Please add DOI's to the references in the paper, if references have a DOI.
