### 2018-08-05 andrewthomasjones/logKDE#2 @strengejacke
[JOSS-review] Add automated test-suite
For https://github.com/openjournals/joss-reviews/issues/870

The repo is missing an automated test-suite, which tests if functions are working as intended (e.g. using _testthat_ package). Please add tests that at least check if function arguments work w/o error, if possible, also check return results to match expected return values.
