### 2018-08-05 #870 @strengejacke
The subsmission (i.e. repo and paper) is well prepared, there's actually not much to complain about. I have a few minor issues you may consider. Since the package is "small" and compact, focussing on few functions about kernel density estimation, I could review this submission rather quickly. :-)
