### 2018-08-06 243f1da
References are added to manual
