### 2018-08-06 68e7657
References are added to manual
