### 2018-08-06 457b2ae
Added DOIs
