+++ paper.bib
@@ -13,6 +13,7 @@
   volume={27},
   pages={832--837},
   year={1956},
+  url={https://www.doi.org/10.1214/aoms/1177728190}
 }
 
 @article{parzen1962estimation,
@@ -22,6 +23,7 @@
   volume={33},
   pages={1065--1076},
   year={1962},
+  url={https://www.doi.org/10.1214/aoms/1177704472}
 }
 
 @book{silverman2018density,
@@ -55,6 +57,7 @@
   year={2002},
   location={New York},
   publisher={Springer},
+  url={https://doi.org/10.1007/978-0-387-21706-2}
 }
 
 @book{keen2010graphics,
@@ -70,7 +73,8 @@
   author={Maindonald, J and Braun, J},
   location={Cambridge},
   year={2010},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  url={https://doi.org/10.1017/CBO9781139194648}
 }
 
 @book{verzani2014using,
@@ -95,7 +99,8 @@
   journal={L'Actualite economique},
   volume={91},
   pages={141--159},
-  year={2015}
+  year={2015},
+  url={https://www.doi.org/10.7202/1036917ar}
 }
 
 @book{eddelbuettel2013seamless,
@@ -103,7 +108,8 @@
   author={Eddelbuettel, D},
   year={2013},
   location={New York},
-  publisher={Springer}
+  publisher={Springer},
+  url={https://www.doi.org/10.1007/978-1-4614-6868-4}
 }
 
 @article{RJ2016045,
@@ -124,5 +130,6 @@
   journal={Journal of Statistical Software},
   volume={56},
   pages={1--30},
-  year={2018}
+  year={2018},
+  url={https://www.doi.org/10.18637/jss.v084.i05}
 }
