### 2018-08-06 andrewthomasjones/logKDE#1 @hiendn
Hi @strengejacke,

All DOIs that could be found have now been added to the paper. Please see the latest generated PDF document over at [https://github.com/openjournals/joss-reviews/issues/870](https://github.com/openjournals/joss-reviews/issues/870).

Please let me know if you are happy with this.
