### 2018-08-06 0acbfba
Test functions testthat
