+++ DESCRIPTION
@@ -14,5 +14,6 @@ LinkingTo: Rcpp
 Imports: Rcpp, pracma
 SystemRequirements: C++11
 RoxygenNote: 6.0.1
-Suggests: R.rsp
+Suggests: R.rsp,
+    testthat
 VignetteBuilder: R.rsp
