+++ tests/testthat/test-logdensity.R
@@ -17,7 +17,7 @@ test_that("Check that the default settings of logdensity generates a 'density' o
 test_that("Check that different kernel inputs for logdensity generates a 'density' object.", {
 
   # Generate some random data
-  SAMPLE1000 <- rchisq(100,10)
+  SAMPLE100 <- rchisq(100,10)
 
   # Check that logdensity produces a 'density' object for the kernel
   expect_is(logdensity(SAMPLE100,kernel = 'epanechnikov'), 'density')
