### 2018-08-06 andrewthomasjones/logKDE#2 @hiendn
Hi @strengejacke,

As you have recommended, we have used the `testthat` package to construct a suite of testing functions, that are now contained in the [tests](https://github.com/andrewthomasjones/logKDE/tree/master/tests) folder of the repository. The suite tests the four main functions of the package over a variety of input combinations in order to assess that the functions produce the correct output classes and values, without error.

A run of the test suite is included at the bottom of the [readme.Rmd](https://github.com/andrewthomasjones/logKDE/blob/master/README.Rmd) document. Please let me know if you believe that these tests are sufficient.
