### 2018-08-06 #870 @hiendn
Hi @strengejacke, I believe that we have resolved all three of the issues that you had raised on the [package repo](https://github.com/andrewthomasjones/logKDE). Please let me know if you are happy with the changes that have been made. 😄

I have left specific responses regarding each of your comments in the respective issues pages.
