### 2018-08-06 andrewthomasjones/logKDE#1 @strengejacke
Thanks, looks fine!
