### 2018-08-06 #870 @strengejacke
@arfon I have finished my review. All raised issues have been addressed adequately, so I have nothing more to complain about and I would recommend accepting the paper.
