### 2018-08-06 #870 @hiendn
@strengejacke - Thank you very much for your quick and thorough review.

@arfon - The Zenodo DOI of the archive for the reviewed version of the package is 10.5281/zenodo.1339352. I have also edited the paper.md file to now state the address of the latest Zenodo-archived release. Thank you for your editorial work!
